[enh] is_valid_lang moved to utils
